Everthing works except the test for dividing by zero. But if I try to divide by zero in the browser it properly shows my error message

+++ app/controllers/calculator_controller.rb
@@ -22,7 +22,7 @@ class CalculatorController < ApplicationController
 
   def divide
     if (params[:d_second].to_i != 0) && (params[:d_first] && params[:d_second])
-      @d_answer = params[:d_first].to_i / params[:d_second].to_i
+      @d_answer = params[:d_first].to_f / params[:d_second].to_f
       render :divide
     elsif params[:d_second].to_i == 0
       @d_answer = 'Error: you cannot divide by zero.'
